+++ YNDropDownMenu/Classes/YNDropDownView.swift
@@ -18,10 +18,6 @@ open class YNDropDownView: UIView {
         self.delegate?.hideMenu()
     }
     
-    @available(*, deprecated, message: "use changeMenu(title: String, at index: Int) instead")
-    open func changeMenuTitleAt(index: Int, title: String) {
-        self.delegate?.changeMenu(title: title, at: index)
-    }
     
     open func changeMenu(title: String, at index: Int) {
         self.delegate?.changeMenu(title: title, at: index)